Refactor ExpressionExtensions and tests.
