+++ src/MudBlazor.UnitTests/Extensions/ExpressionExtensionsTests.cs
@@ -1,15 +1,17 @@
 ﻿using FluentAssertions;
-using MudBlazor.Utilities;
+using MudBlazor;
 using NUnit.Framework;
 using System.Linq.Expressions;
 using System;
+using System.ComponentModel.DataAnnotations;
 
-namespace UtilityTests
+namespace MudBlazor.UnitTests.Extensions
 {
     public class ExpressionExtensionsTests
     {
         private class TestClass
         {
+            [Display(Name = "Display Attribute")]
             public string Field1 { get; set; }
 
             public TestClass1 TestClass1 { get; set; }
@@ -31,7 +33,7 @@ namespace UtilityTests
         public void GetFullPathOfMemberTest()
         {
             var model = new TestClass();
-            
+
             Expression<Func<string>> expression = () => model.Field1;
 
             expression.GetFullPathOfMember().Should().Be("Field1");
@@ -55,6 +57,26 @@ namespace UtilityTests
             Expression<Func<string>> expression = () => model.TestClass1.TestClass2.Field1;
 
             expression.GetFullPathOfMember().Should().Be("TestClass1.TestClass2.Field1");
+        }
+
+        [Test]
+        public void GetDisplayNameStringTest1()
+        {
+            var model = new TestClass();
+
+            Expression<Func<string>> expression = () => model.Field1;
+
+            expression.GetDisplayNameString().Should().Be("Display Attribute");
+        }
+
+        [Test]
+        public void GetDisplayNameStringTest2()
+        {
+            var model = new TestClass1();
+
+            Expression<Func<string>> expression = () => model.Field1;
+
+            expression.GetDisplayNameString().Should().Be("");
         }
     }
 }
